requirements.txt
httpx==0.27.0
pyLoraRFM9x==1.0.2
python-dotenv==1.0.1 